requirements.txt
Flask==1.1.1
PyYAML==5.1.2
py-ztj-registry==0.0.2
